requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1

# Database
sqlmodel==0.0.22
psycopg2-binary==2.9.10
supabase>=2.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic[email]==2.10.4
email-validator==2.2.0

# AI Integration
google-generativeai==0.8.3
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.154.0

# Firebase (Notifications)
firebase-admin==6.6.0

# WebSocket
websockets==14.1

# HTTP Client
httpx==0.28.1